pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plotql"
version = "0.1.0"
description = "SQL-like DSL for terminal plotting with Polars and Textual"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "polars>=0.20.0",
    "textual>=0.40.0",
    "textual-image[textual]>=0.6.0",
    "textual-autocomplete>=3.0.0a0",
    "matplotlib>=3.8.0",
    "Pillow>=10.0.0",
    "pyperclip>=1.8.0",
]

[project.scripts]
plotql = "plotql.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
